requirements.txt
# Core Dependencies
strands-agents[openai,a2a]>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv==1.0.1

# Web Framework
flask==3.0.2
flask-cors==4.0.0

# HTTP Requests
requests==2.31.0
httpx>=0.28.1,<1.0.0

# Data Handling
pydantic>=2.11.3,<3.0.0

# Date/Time Handling
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Development
black==24.2.0
flake8==7.0.0
